refactor: Extract codes related to constructing the context for searching
