+++ src/commands/select.rs
@@ -1,8 +1,7 @@
-use globset::GlobBuilder;
-
-use crate::config::option::{CaseSensitivity, SelectOption};
+use crate::config::option::SelectOption;
 use crate::context::AppContext;
 use crate::error::JoshutoResult;
+use crate::util::search::SearchContext;
 
 use super::cursor_move;
 
@@ -55,28 +54,12 @@ fn select_with_pattern(
     options: &SelectOption,
 ) -> JoshutoResult {
     let case_sensitivity = context.config_ref().search_options_ref().case_sensitivity;
-    let pattern_lower = pattern.to_lowercase();
+    let search_context = SearchContext::new_glob(pattern, case_sensitivity)?;
 
-    let (pattern, actual_case_sensitivity) = match case_sensitivity {
-        CaseSensitivity::Insensitive => (pattern_lower.as_str(), CaseSensitivity::Insensitive),
-        CaseSensitivity::Sensitive => (pattern, CaseSensitivity::Sensitive),
-        // Determine the actual case sensitivity by whether an uppercase letter occurs.
-        CaseSensitivity::Smart => {
-            if pattern_lower == pattern {
-                (pattern_lower.as_str(), CaseSensitivity::Insensitive)
-            } else {
-                (pattern, CaseSensitivity::Sensitive)
-            }
-        }
+    let glob = match &search_context {
+        SearchContext::Glob(glob) => glob,
+        _ => unreachable!(),
     };
-
-    let glob = GlobBuilder::new(pattern)
-        .case_insensitive(matches!(
-            actual_case_sensitivity,
-            CaseSensitivity::Insensitive
-        ))
-        .build()?
-        .compile_matcher();
 
     if let Some(curr_list) = context.tab_context_mut().curr_tab_mut().curr_list_mut() {
         let mut found = 0;
